Clean up selection model properly in completion.
http://qt-project.org/doc/qt-5/qabstractitemview.html#setModel :

This function will create and set a new selection model, replacing any model
that was previously set with setSelectionModel(). However, the old selection
model will not be deleted as it may be shared between several views. We
recommend that you delete the old selection model if it is no longer required.