Desktopy goodness
.desktop file, icon, and a note about installing them, for packagers.

Please, for the love of flying spaghetti monster, someone make a better icon.
This is just here so the desktop file has something to use.

.desktop file worked on kde4 and kde5. Should work on everything else too as
it's dead simple

Icon is from breeze icons licensed under the LGPL
https://github.com/NitruxSA/plasma-next-icons/blob/master/apps/software/network-manager.svg

Add a note for packagers since we can't use setup.py to install non-code
resources to system locations.