Clean up standarddir handling #2.
We already attempted this in c5a2039da431caa03b839e785d4046f05492c84d, but
having the directories as module attributes means they'll be created on start
(rather than when they're actually used), and it'd also be impossible to change
them after init for some reason in the future.

To still have a nice short API, we simply change the attributes to functions.