More user-friendly link text matching.
Allow the user to type any part of the link text when filtering down the
links, not just the start of the text. The start is often obscured by a
hint, or begins with a number making it impossible to select.

Also allow for uppercase characters when filtering. Previously any
uppercase character would abort matching.