Add a command to show the keystring for a key.
This fixes at least part of #658. Note that this is not the same as <C-v> in
Vim. In Vim, <C-v> is used to insert keys (not show their keystring). You can do
cool stuff like <C-v>x07 to insert 0x07, or `<C-v>u2713` to insert ✓

Adding that to the input bar and/or insert mode in text forms is another thing
altogether...

I've mapped in to <C-k> by the way, as <C-v> is already taken...