Adds completion for tab-focus.
This is so you can jump between tabs based on tab title or URL instead
of just index just like the `buffer` command in vim. Adds a `b` alias
for `tab-focus` to be more like vim too.

The method of updating the completion model (rebuild the whole thing
every time a tab is created, removed or updated) is inefficient but is
still very fast for even unreasonable numbers of open tabs.