Add per-domain referer policy implementation.
Probably needs a bit of refactoring with the multiple return points and not
checking for url validity.

Would prefer to base the decision whether to send the referer header on the
rquest url not the current one but this is easier for the "make things on this
page work" use case. Basing is on destination url wold require a ui upgrade to
see what was going on.