Fix for issue #1596 - Make emergency backup
When editing a form field or similar in an external editor (Ctrl+E) and the
corresponding field vanishes, then the edited text will be lost. This fixes the
problem by creating a more permanent temp file in such cases and telling the
user about it, so they can manually restore the lost contents if needed.

Note that this will spam your /tmp (or whereever your system temp is) if it
happens a lot, since those files will never be cleaned up by qutebrowser.