initial implementation of qutewm
qutewm should be the window manager that we can use in tests running
in Xvfb (or on travis). Some tests are currently not possible because
they have to deal with multiple windows and focusing the right one.

qutewm aims to solve that by implementing a basic subset of ICCCM
(Inter-Client Communication Conventions Manual) and EWMH (Extended
Window Manager Hints), allowing us to use either tools like xdotool or
built-in functions to deal with window focus.

Note that qutewm is not meant as a standalone window manager. It's a
test tool, nothing more.