qutewm: also use SubstructureRedirectMask
This is actually the mask that makes a WM a WM, as only one listener is
allowed per window. So if another WM is running, we will generate a X
error, which we can later catch and exit.

Unfortunately, now we're intercepting some necessary events, which means
we need to re-send them, which adds some boilerplate.