utilcmds: add :debug-focus-window
The command uses QApplication.alert, which in turn sets the urgency hint
on the given window.

This will only work in qutewm, as other window managers will most likely
just flash the window bar or something else, as no one really wants the
window to switch on urgency. However, this feature is useful for tests
and does not require external libraries like python-xlib or python-ewmh,
as Qt takes care of everything.

The optimal thing of course would be to use the _NET_ACTIVE_WINDOW
client message, but Qt does not use this (not even for
QWidget.activateWindow)