qutewm: start only once per session
I noticed that it doesn't make sense for qutewm to start for a single
test, especially if the qutebrowser window is already created, as that
would mean it's not managed by qutewm. I *think* this led to some
strange X errors about BadWindow, because the window manager was somehow
gone after the test, but I'm not sure...

A window manager should be something that starts before the application
and quits after the application quits. This patch changes the behaviour
of qutewm to match that:

The "qutewm" fixture makes sure a qutewm instance is running in the
session. It saves the manager as request.session._qutewm, so that other
fixtures can retrieve it. "quteproc_process" and "quteproc_new" have
been changed to require "qutewm", to make sure the WM is started before
qutebrowser. If the manager can't be started, "qutewm" does nothing. The
tests that don't require a WM should run fine, as they didn't have a WM
before anyway (as we found out, herbstluftwm was useless).

The "qutewm_manager" fixture is on autouse and does nothing for tests
that don't have the @qutewm mark. For @qutewm marked tests, it skips
them if qutewm is not running (either because we're not on Linux or
because another WM is running). It also calls the .before_test and
.after_test methods of the testprocess and makes sure the log is
captured and shown. It's on autouse because we mainly need it in BDD
feature tests, where a @qutewm marker is replacing the old @no_xvfb
marker.

Altogether this should make sure that the WM is started once per test
session (just like Xvfb is started once), and shouldn't influence tests
where the WM is not needed.