qutewm: more testing and fixed events
First, the event handling is fixed. The log line said "activated", not
"focused", so the event wasn't actually triggered. This has been
changed.

Also, to really make sure the "activated" message is not being send too
early, it's now sent on the FocusIn event, not on the call to .activate.
This makes sure that the focus really has changed, as before the first
thing we did was to send the message.

Then this adds another test for qutewm, which makes sure window
switching via urgency hinting works.

All tests now wait for some milliseconds. This is to make sure that
after the qutewm.window_focused signal has been fired, Qt can also
register the focus change. So we enter the event loop and quit it
shortly afterwards. If we don't do that, the activeWindow method may not
return the correct result.