testprocess: use QProcess.kill on Windows
webserver_sub was still lingering around after tox has finished and it
blocked the Appveyor build for up to an hour. This patch should make
sure that even on Windows, webserver_sub is "properly" killed.

Based on some quick tests, it seems that the SIGTERM handler is ignored
anyway, so it's probably okay to just kill the child.