fix handling of choices with no type given
Fixes #1871.

The fix is basically the one of @The-Compiler with a small bugfix. The
value that caused trouble was the *args one, as that was a list and not
a string. This patch makes that *args and **kwargs are returned
unchanged, which means the fix as The-Compiler wrote it works.